requirements.txt
Django==2.2
djangorestframework==3.9.2
PyJWT==1.7.1
pytz==2019.1
requests==2.21.0
urllib3==1.25.8
